api/requirements.txt
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
argon2-cffi>=23.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
